ml-service/requirements.txt
py -m pip install -r requirements.txtflask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
joblib==1.3.2
